requirements.txt
anyio==3.6.1
attrs==22.1.0
click==8.1.3
fastapi==0.79.0
h11==0.13.0
idna==3.3
iniconfig==1.1.1
loguru==0.6.0
numpy==1.23.1
opencv-python==4.6.0.66
packaging==21.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pycodestyle==2.9.1
pydantic==1.9.2
pyparsing==3.0.9
pystun3==1.0.0
pytest==7.1.2
pyusb==1.2.1
sh==1.14.3
sniffio==1.2.0
SQLAlchemy==1.4.40
starlette==0.19.1
tomli==2.0.1
typing-extensions==4.3.0
uvicorn==0.18.2
